+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ## Changelog
+
+- Add Tooltips to a bunch of Encounter Alert options